Handle FSIARGS correctly in zsh
In zsh, a variable with a space in it will be treated as if it was
quoted, which is not how bash behaves. There's a zsh-specific option you
can set to consider it as two words instead, but that's not friendly to
strictly-POSIX-compliant shells like dash. Instead, for maximum
compatibility and simplicity, simply split it into two variables.