Check that F# is installed in init.fsx
If F# is not installed (say a Linux user has installed the mono-complete
package thinking that F# was included in it), the error message that the
user will see when he runs build.sh or build.cmd is "Target named
'Rebuild' not found in the project". Instead, we should check (once) in
init.fsx to make sure that F# is properly installed, and give a clearer
error message if it isn't.

Fixes #253.